refactor(cli): move the wizard's channel cluster into its own module
Second cut of the onboard split: the Step3 chat-channel cluster (~580
lines, 15 symbols) moves to onboard_channels.py under the same
discipline as the everos cut -- shared wizard state stays in
onboard_commands, both sides reach across through module references
with call-time attribute access, and every migrated patch target was
mutation-checked one by one (12/12 red against the old module). Also
renames _step5_import to _step6_import: its own step header already
said 6, and the misnamed function sat next to the real _step5 (deep
research), waiting to misdirect the next split.

onboard_commands.py is now 3000 lines, down from 5069.
